pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sd-dmfa"
version = "0.9.0"
description = "Template for coupled System Dynamics (BPTK-Py) and dynamic MFA (flodym)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "flodym>=0.5.0",
  "bptk-py>=2.0.0",
  "rich>=13.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "ruff>=0.4",
  "mypy>=1.8",
  "pre-commit>=3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
